+++ docs/helpers/request.md
@@ -1,4 +1,4 @@
-# apply.js
+# request.js
 
 ## queryObjectToString
