+++ src/helpers/main.js
@@ -1,6 +1,6 @@
 export * from './form';
 export * from './jsQuery';
-export * from './url';
+export * from './request';
 export * from './apply';
 export * from './time';
 export * from './Either';
